Новая папка 2/requirements.txt
telebot~=0.0.5
python-dotenv~=1.0.1
openai~=1.30.5